pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentic-rag-for-dummies"
version = "0.1.0"
description = "A modular Agentic RAG application built with LangGraph, Gradio, and Qdrant."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
  { name = "Agentic RAG Maintainers" }
]
dependencies = [
  "docling",
  "fastembed==0.7.4",
  "gradio==6.3.0",
  "Jinja2==3.1.6",
  "langchain==1.2.3",
  "langchain-core==1.2.7",
  "langchain-huggingface==1.2.0",
  "langchain-openai==1.0.1",
  "langchain-ollama==1.0.1",
  "langchain-qdrant==1.1.0",
  "langchain-text-splitters==1.1.0",
  "langfuse==4.0.0",
  "langgraph==1.0.5",
  "langgraph-checkpoint==3.0.1",
  "langgraph-prebuilt==1.0.5",
  "langgraph-sdk==0.3.2",
  "pydantic==2.12.5",
  "portalocker>=2.8",
  "PyMuPDF==1.26.6",
  "pymupdf-layout==1.26.6",
  "pymupdf4llm==0.2.9",
  "python-dotenv==1.2.2",
  "qdrant-client==1.16.2",
  "sentence-transformers==5.2.0",
  "tiktoken==0.12.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
]
eval = [
  "Jinja2==3.1.6",
]

[project.scripts]
agentic-rag = "agentic_rag.cli:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]
